+++ scripts/state/game_state.gd
@@ -5,6 +5,7 @@ enum Faction { HERO, BOSS }
 
 var selected_faction: Faction = Faction.HERO
 var selected_character: String = "Warrior"
+var last_selected_hero: String = "Warrior"
 var is_game_over: bool = false
 var victory_type: VictoryType = VictoryType.NONE
 
@@ -15,5 +16,6 @@ func end_game(type: VictoryType):
 func reset():
 	selected_faction = Faction.HERO
 	selected_character = "Warrior"
+	last_selected_hero = "Warrior"
 	is_game_over = false
 	victory_type = VictoryType.NONE